+++ mobile/android/android-components/docs/changelog.md
@@ -39,6 +39,10 @@ permalink: /changelog/
 * **feature-sitepermissions**
   *  Fixed [#6322](https://github.com/mozilla-mobile/android-components/issues/6322), now  `SitePermissionsStorage` allows to indicate an optional reference to `DataCleanable`.
 
+* **browser-menu**
+  * Added `canPropagate` param to all `BrowserMenuHighlight`s, making it optional to be displayed in other components
+  * Changed `BrowserMenuItem.getHighlight` to filter the highlights which should not propagate.
+
 # 36.0.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v35.0.0...v36.0.0)